feat: realm status
