+++ src/components/cityview/realm/RealmInfoComponent.tsx
@@ -1,29 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { ResourceIcon } from '../../../elements/ResourceIcon';
 import { findResourceById } from '../../../constants/resources';
+import { ReactComponent as ShieldIcon } from '../../../assets/icons/common/shield.svg';
 
 type RealmInfoComponentProps = {}
 
 export const RealmInfoComponent = ({ }: RealmInfoComponentProps) => {
-
-    const dummyResources = [
-        {
-            resourceId: 3,
-            amount: 150
-        },
-        {
-            resourceId: 4,
-            amount: 32
-        },
-        {
-            resourceId: 18,
-            amount: 120,
-        },
-        {
-            resourceId: 20,
-            amount: 4
-        }
-    ]
 
     return (
         <>
@@ -38,6 +20,10 @@ export const RealmInfoComponent = ({ }: RealmInfoComponentProps) => {
                     <div className="text-xxs text-gold">
                         0x...loaf
                     </div>
+                </div>
+                <div className='flex items-center justify-center px-3 py-2 ml-auto text-xs text-white rounded-full cursor-default bg-light-red'>
+                    <ShieldIcon className="mr-2 fill-white" />
+                    <span>Vulnerable</span>
                 </div>
             </div>
         </>
