+++ src/elements/tab/tabs.tsx
@@ -18,8 +18,8 @@ export const VARIANTS = {
   },
   primary: {
     tab: {
-      base: 'relative mx-1 z-10  border-t transition-all duration-200 border-x border-transparent !outline-none -mb-[1px] text-xs py-3 px-2',
-      active: 'rounded-t-xl !border-gold text-gold bg-gradient-to-b from-black to-[#151515]',
+      base: 'relative mx-1 z-10 rounded-t-xl  border-t transition-all duration-200 border-x border-transparent !outline-none -mb-[1px] text-xs py-3 px-2',
+      active: '!border-gold text-gold bg-gradient-to-b from-black to-[#151515]',
       inactive: 'text-gray-gold',
     },
     tabList: 'flex w-full',
