+++ src/modules/RealmManagementModule.tsx
@@ -5,7 +5,7 @@ import { BaseContainer } from "../containers/BaseContainer"
 
 const RealmManagementModule = () => {
     return (
-        <BaseContainer className="h-full p-0 rounded-t-3xl">
+        <BaseContainer className="h-full !p-0 rounded-t-3xl">
             <RealmInfoComponent />
             {/* <RealmStatusComponent /> */}
             <RealmManagementComponent />
